Day 3 part 2
In all fairness, we got some help from Reddit. The brute force combinatorial approach wasn't working. But because the numbers were ordered, we could just simply build it by finding the biggest number possible at the beginning of the string, and then do the same for substrings (leaving a minimum of 12 so we can find all).